pyproject.toml
[tool.poetry]
name = "Printix_Server_Automation"
version = "0.1.0"
description = "\"Printix Server Automation\""
authors = ["RameshKunamneni <ramesh.kunamneni@printix.net>"]
readme = "README.md"
packages = [{include = "pyproject"}]

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.2.1"
allure-pytest = "^2.12.0"
openapi = "^1.1.0"
python-openapi = "^0.1.8"
requests = "^2.28.2"
logging = "^0.4.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"